requirements.txt
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
bcrypt==3.1.4
beautifulsoup4==4.6.0
blinker==1.4
bzr==2.7.0.dev1
cffi==1.11.5
chardet==2.0.1
charmtools==1.0.0
Cheetah==2.4.4
click==6.7
colorama==0.2.5
command-not-found==0.3
configobj==4.7.2
configparser==3.5.0
cryptography==2.2.2
debtagshw==0.1
defer==1.0.6
dirspec==13.10
duplicity==0.6.23
EasyProcess==0.2.3
enum34==1.1.6
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
futures==3.2.0
html5lib==0.999
httplib2==0.8
idna==2.7
ipaddress==1.0.22
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
lockfile==0.8
lxml==3.3.3
Markdown==2.4
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.8.2
oauth==1.0.1
oauthlib==0.6.1
oneconf==0.3.7.14.4.1
PAM==0.4.2
paramiko==1.10.1
passlib==1.7.1
pbr==4.2.0
pexpect==3.1
Pillow==2.3.0
piston-mini-client==0.7.5
pycparser==2.18
pycrypto==2.6.1
pycups==1.9.66
Pygments==1.6
pygobject==3.12.0
pylint==1.9.3
PyMySQL==0.9.0
pyOpenSSL==0.13
pyserial==2.6
pysmbc==1.0.14.1
python-apt===0.9.3.5ubuntu3
python-debian===0.1.21-nmu2ubuntu2
PyVirtualDisplay==0.2.1
pyxdg==0.25
PyYAML==3.10
reportlab==3.0
requests==2.2.1
SecretStorage==2.0.0
selenium==3.14.0
sessioninstaller==0.0.0
simplejson==3.3.1
singledispatch==3.4.0.3
six==1.5.2
software-center-aptd-plugins==0.0.0
SQLAlchemy==1.2.10
ssh-import-id==3.21
system-service==0.1.6
tbselenium==0.3.3
testresources==0.2.7
Twisted-Core==13.2.0
Twisted-Web==13.2.0
unity-lens-photos==1.0
urllib3===dev
virtualenv==16.0.0
virtualenv-clone==0.3.0
wadllib==1.3.2
Werkzeug==0.14.1
wrapt==1.10.11
WSGIserver==1.3
WTForms==2.2.1
xdiagnose===3.6.3build2
zope.interface==4.0.5
